requirements.txt
google-generativeai
fastapi
uvicorn
pydantic
python-dotenv
textblob
